Initial step and first wip refactoring of wizard
