+++ app/views/admin/participants/change_relationship/reason_for_change.html.erb
@@ -1,22 +1,22 @@
-<% title = "Which type of appropriate body have you appointed?" %>
+<% title = "Why are you changing this participant’s training programme?" %>
+
 <% content_for :title, title %>
 
 <% content_for :before_content, govuk_back_link(text: "Back", href: wizard_back_link_path) %>
 
 <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
+        <span class="govuk-caption-xl"><%= @wizard.participant_role_and_name %></span>
+
         <%= form_with model: @wizard.form, url: url_for(action: :update), scope: @wizard.session_key, method: :put do |f| %>
             <%= f.govuk_error_summary %>
+
             <%= f.govuk_collection_radio_buttons(
-                :appropriate_body_type,
-                @wizard.form.choices,
+                :reason_for_change,
+                @wizard.form.options,
                 :id,
                 :name,
-                legend: { text: title, tag: :h1, size: 'xl' },
-                caption: { text: @school.name, size: 'xl' }) do %>
-
-                <p class="govuk-body">If your school works with more than one, first tell us about the appropriate body appointed for the majority of your ECTs. You can add the others later.</p>
-            <% end %>
+                legend: { text: title, tag: 'h1', size: 'xl' }) %>
 
             <%= f.govuk_submit "Continue" %>
         <% end %>
