+++ app/views/schools/cohort_setup/appropriate_body_appointed.html.erb
@@ -5,7 +5,7 @@
 
 <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
-        <%= form_with model: @wizard.form, url: url_for(action: :update), scope: @wizard.to_key, method: :put do |f| %>
+        <%= form_with model: @wizard.form, url: url_for(action: :update), scope: @wizard.session_key, method: :put do |f| %>
             <%= f.govuk_error_summary %>
             <%= f.govuk_collection_radio_buttons(
                 :appropriate_body_appointed,
