+++ app/views/schools/cohort_setup/expect_any_ects.html.erb
@@ -8,7 +8,7 @@
     <div class="govuk-grid-column-two-thirds">
         <span class="govuk-caption-xl"><%= @wizard.school.name %></span>
 
-        <%= form_with model: @wizard.form, url: url_for(action: :update), scope: @wizard.to_key, method: :put do |f| %>
+        <%= form_with model: @wizard.form, url: url_for(action: :update), scope: @wizard.session_key, method: :put do |f| %>
             <%= f.govuk_error_summary %>
 
             <%= f.govuk_collection_radio_buttons(
