+++ app/views/schools/cohort_setup/providers_relationship_has_changed.html.erb
@@ -15,7 +15,7 @@
         <p class="govuk-body">You can form a new partnership with a lead provider and delivery partner, or
             change the training programme.</p>
 
-        <%= form_with model: @wizard.form, url: url_for(action: :update), scope: @wizard.to_key, method: :put do |f| %>
+        <%= form_with model: @wizard.form, url: url_for(action: :update), scope: @wizard.session_key, method: :put do |f| %>
             <%= f.govuk_submit "Confirm" %>
         <% end %>
     </div>
