+++ app/views/schools/cohort_setup/what_changes_confirmation.html.erb
@@ -6,7 +6,7 @@
     <div class="govuk-grid-column-two-thirds">
         <%= render partial: @wizard.what_changes + "_confirmation" %>
 
-        <%= form_with model: @wizard.form, url: url_for(action: :update), scope: @wizard.to_key, method: :put do |f| %>
+        <%= form_with model: @wizard.form, url: url_for(action: :update), scope: @wizard.session_key, method: :put do |f| %>
             <%= f.govuk_submit "Confirm" %>
         <% end %>
     </div>
